services/gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "common",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
common = { workspace = true }
